requirements.txt
flask==3.1.0
gunicorn==23.0.0
opencv-python-headless==4.11.0.86
numpy==2.2.4
scipy==1.15.2
Pillow==11.1.0
PyWavelets==1.8.0
scikit-image==0.25.2
qrcode[pil]==8.0
reportlab==4.4.0
yt-dlp==2026.3.17
torch
transformers
